snippet/up-down/requirements.txt
azure-common==1.1.26
azure-core==1.12.0
azure-mgmt-compute==19.0.0
azure-mgmt-core==1.2.2
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.3
